fix tenant creation.

+++ database/migrations/tenant/2026_04_24_100008_add_customer_payment_type_to_customer_payments_table.php
@@ -7,15 +7,19 @@
     return new class extends Migration {
         public function up() : void
         {
-            Schema::table( 'customer_payments' , function (Blueprint $table) {
-                $table->unsignedTinyInteger( 'customer_payment_type' )->nullable()->after( 'payment_method_id' );
-            } );
+            if ( ! Schema::hasColumn( 'customer_payments' , 'customer_payment_type' ) ) {
+                Schema::table( 'customer_payments' , function (Blueprint $table) {
+                    $table->unsignedTinyInteger( 'customer_payment_type' )->nullable()->after( 'payment_method_id' );
+                } );
+            }
         }
 
         public function down() : void
         {
-            Schema::table( 'customer_payments' , function (Blueprint $table) {
-                $table->dropColumn( 'customer_payment_type' );
-            } );
+            if ( Schema::hasColumn( 'customer_payments' , 'customer_payment_type' ) ) {
+                Schema::table( 'customer_payments' , function (Blueprint $table) {
+                    $table->dropColumn( 'customer_payment_type' );
+                } );
+            }
         }
     };
